chore(htn): Share task network implementation between method and initial taks network.
